requirements.txt
anyio==4.9.0
apify_client==1.10.0
apify_shared==1.4.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dateparser==1.2.1
dotenv==0.9.9
filelock==3.18.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.5.1
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.6
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
more-itertools==10.7.0
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
